push ParameterState to step-capture-parameters
